xdiff: use uint64_t instead of long for xdl_bogosqrt()

+++ xdiff/xdiffi.c
@@ -509,7 +509,7 @@ int xdl_do_classic_diff(xdfenv_t *xe, uint64_t flags)
 	kvdf += xe->xdf2.nreff + 1;
 	kvdb += xe->xdf2.nreff + 1;
 
-	xenv.mxcost = xdl_bogosqrt(ndiags);
+	xenv.mxcost = (long)xdl_bogosqrt(ndiags);
 	if (xenv.mxcost < XDL_MAX_COST_MIN)
 		xenv.mxcost = XDL_MAX_COST_MIN;
 	xenv.snake_cnt = XDL_SNAKE_CNT;
